+++ generators/code-test/helpers.js
@@ -1,7 +1,1 @@
-require('sugar');
-
-module.exports = function(Handlebars) {
-  Handlebars.registerHelper('humanize', function(options) {
-    return new Handlebars.SafeString(options.fn(this).humanize());
-  });
-};
+module.exports = require('../helpers');